Delete Function
The delete function wasn't deleting the account once the user chose to close their account. It was also printing that the account was closed even if the password was wrong. Although it isn't completely fixed because the delete function now works but it deletes everything in the index. I haven't figured out a way to delete in a chain yet.